Animations

+++ Course3/ReactNative/confusion2/components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { View, FlatList } from "react-native";
+import { View, FlatList, Text, Alert } from "react-native";
 import { ListItem } from "react-native-elements";
 import { connect } from "react-redux";
 import { baseUrl } from "../shared/baseUrl";
@@ -31,7 +31,26 @@ class Favorites extends Component {
         {
           text: "Delete",
           type: "delete",
-          onPress: () => this.props.deleteFavorite(item.id)
+          onPress: () => {
+            Alert.alert(
+              "Delete Favorite ?",
+              "Are you sure you wish to delete the favorite dish " +
+                item.name +
+                " ?",
+              [
+                {
+                  text: "Cancel",
+                  onPress: () => console.log(item.name + "Not Deleted"),
+                  style: "cancel"
+                },
+                {
+                  text: "OK",
+                  onPress: () => this.props.deleteFavorite(item.id)
+                }
+              ],
+              { cancelable: false }
+            );
+          }
         }
       ];
       return (
